Import changes from SVN into git
This changes primary affects the shard packages.
Only some import deletion in daq-core are effected because of the remove of this classes in the shared packages.

The majority of the changes are adding @Data annotation from Lombok to classes which are serialized because of the server request.
Because of the annotation a lot of setter and getter got removed.
This changes are done because of the Jackson serialiazation.

Furthermore toXML and fromXML methods got removed in all classes which will be serialiazed by Jackson.

In order to generalize requests from the server  and responses from the DAQ new interfaces got added:
- All classes which holds request infomration implement the SertverRequest interface
- All classes which holds DAQ response information implement the DAQResponse interface

The interfaces will help then on the server and the daq to distinguish betweent the different messages.