Share common properties between configuration classes
This commit refactors some `@ConfigurationProperties` classes that had duplicate property names
in different layers (client/server or server/daq) into common classes. Where additional properties
are needed in the same namespace, the common class is simply extended to inherit the common
properties and the additional properties are added to the subclass. This avoids unnecessary
duplication.

This commit also fixes #73.