Remove demo data SQL and corresponding config option
Given the upcoming new Elasticsearch configuration index, it is no longer
feasible to provide a demo configuration by simply inserting tags into the
DB. This is due to the fact that at server startup, those tags will be
loaded into the cache, but not the configuration index, and hence will not
be searchable.

The proposed solution to provide a demo configuration is to use the client
API directly from the hostmetrics (and testhandler) DAQ to automatically
configure its own tags by itself.