Resolve ambiguity for mybatis Java type column definitions
The `ObjectTypeHandler` class is responsible for converting Object attributes of mapped
entities to/from JSON (e.g. the tag value). When using the new `@MappedTypes` annotation
on this class and specifying `Object.class` as a mapped type, mybatis was then using this
type handler for all columns that did not have an explicit Java type declared. This means
that JSON serialisation/deserialisation was being attempted for incorrect columns.

This MR adds explicit Java type declarations for all columns that did not previously have
them, to prevent mybatis from using the `ObjectTypeHandler` for those columns.