Put features and use-cases in index, fix links
Remove the features and use-cases as separate pages and
instead append them to the root/index file. This is the
original behaviour in c2mon.web.cern.ch/c2mon/
Fix the links in the menu.yml file to point to the new
headers. Adapt the slugify filter to let '#' characters
through in order for the href links to work properly